requirements.txt
gast==0.1.3
nbconvert==5.2.1
nbformat==4.3.0
pandas==0.20.2
requests==2.9.1
PyGithub==1.35
